+++ src/amplitude-snippet.js
@@ -3,7 +3,7 @@
   var as = document.createElement('script');
   as.type = 'text/javascript';
   as.async = true;
-  as.src = 'https://d24n15hnbwhuhn.cloudfront.net/libs/amplitude-2.6.2-min.gz.js';
+  as.src = 'https://d24n15hnbwhuhn.cloudfront.net/libs/amplitude-2.7.0-min.gz.js';
   as.onload = function() { window.amplitude.runQueuedFunctions(); };
   var s = document.getElementsByTagName('script')[0];
   s.parentNode.insertBefore(as, s);
@@ -13,7 +13,7 @@
       this._q.push([fn].concat(Array.prototype.slice.call(arguments, 0))); return this;
     };
   }
-  var identifyFuncs = ['add', 'set', 'setOnce', 'unset'];
+  var identifyFuncs = ['add', 'append', 'set', 'setOnce', 'unset'];
   for (var i = 0; i < identifyFuncs.length; i++) {
     proxyIdentify(identifyFuncs[i]);
   }