+++ src/amplitude.js
@@ -1,4 +1,4 @@
-var Cookie = require('./cookie');
+var cookieStorage = require('./cookiestorage');
 var JSON = require('json'); // jshint ignore:line
 var language = require('./language');
 var localStorage = require('./localstorage');  // jshint ignore:line
@@ -59,6 +59,7 @@ var Amplitude = function() {
   this._unsentIdentifys = [];
   this._ua = new UAParser(navigator.userAgent).getResult();
   this.options = object.merge({}, DEFAULT_OPTIONS);
+  this.cookieStorage = new cookieStorage().getStorage();
   this._q = []; // queue for proxied functions before script load
 };
 
@@ -110,11 +111,11 @@ Amplitude.prototype.init = function(apiKey, opt_userId, opt_config, callback) {
       this.options.eventUploadPeriodMillis = opt_config.eventUploadPeriodMillis || this.options.eventUploadPeriodMillis;
     }
 
-    Cookie.options({
+    this.cookieStorage.options({
       expirationDays: this.options.cookieExpiration,
       domain: this.options.domain
     });
-    this.options.domain = Cookie.options().domain;
+    this.options.domain = this.cookieStorage.options().domain;
 
     _migrateLocalStorageDataToCookie(this);
     _loadCookieData(this);
@@ -123,7 +124,20 @@ Amplitude.prototype.init = function(apiKey, opt_userId, opt_config, callback) {
         opt_config.deviceId !== null && opt_config.deviceId) ||
         this.options.deviceId || UUID();
     this.options.userId = (opt_userId !== undefined && opt_userId !== null && opt_userId) || this.options.userId || null;
+
+    this._lastEventTime = this._lastEventTime || parseInt(localStorage.getItem(LocalStorageKeys.LAST_EVENT_TIME)) || null;
+    this._sessionId = this._sessionId || parseInt(localStorage.getItem(LocalStorageKeys.SESSION_ID)) || null;
+    this._eventId = this._eventId || parseInt(localStorage.getItem(LocalStorageKeys.LAST_EVENT_ID)) || 0;
+    this._identifyId = this._identifyId || parseInt(localStorage.getItem(LocalStorageKeys.LAST_IDENTIFY_ID)) || 0;
+    this._sequenceNumber = this._sequenceNumber || parseInt(localStorage.getItem(LocalStorageKeys.LAST_SEQUENCE_NUMBER)) || 0;
+    var now = new Date().getTime();
+    if (!this._sessionId || !this._lastEventTime || now - this._lastEventTime > this.options.sessionTimeout) {
+      this._newSession = true;
+      this._sessionId = now;
+    }
+    this._lastEventTime = now;
     _saveCookieData(this);
+    _clearSessionAndEventTrackingFromLocalStorage();
 
     //log('initialized with apiKey=' + apiKey);
     //opt_userId !== undefined && opt_userId !== null && log('initialized with userId=' + opt_userId);
@@ -142,20 +156,6 @@ Amplitude.prototype.init = function(apiKey, opt_userId, opt_config, callback) {
     if (this.options.includeReferrer) {
       this._saveReferrer(this._getReferrer());
     }
-
-    this._lastEventTime = parseInt(localStorage.getItem(LocalStorageKeys.LAST_EVENT_TIME)) || null;
-    this._sessionId = parseInt(localStorage.getItem(LocalStorageKeys.SESSION_ID)) || null;
-    this._eventId = localStorage.getItem(LocalStorageKeys.LAST_EVENT_ID) || 0;
-    this._identifyId = localStorage.getItem(LocalStorageKeys.LAST_IDENTIFY_ID) || 0;
-    this._sequenceNumber = localStorage.getItem(LocalStorageKeys.LAST_SEQUENCE_NUMBER) || 0;
-    var now = new Date().getTime();
-    if (!this._sessionId || !this._lastEventTime || now - this._lastEventTime > this.options.sessionTimeout) {
-      this._newSession = true;
-      this._sessionId = now;
-      localStorage.setItem(LocalStorageKeys.SESSION_ID, this._sessionId);
-    }
-    this._lastEventTime = now;
-    localStorage.setItem(LocalStorageKeys.LAST_EVENT_TIME, this._lastEventTime);
   } catch (e) {
     log(e);
   }
@@ -188,6 +188,10 @@ Amplitude.prototype._loadSavedUnsentEvents = function(unsentKey, queue) {
 
 Amplitude.prototype.isNewSession = function() {
   return this._newSession;
+};
+
+Amplitude.prototype.getSessionId = function() {
+  return this._sessionId;
 };
 
 Amplitude.prototype.nextEventId = function() {
@@ -239,8 +243,9 @@ Amplitude.prototype._sendEventsIfReady = function(callback) {
   return false;
 };
 
+// backwards migration fix for users that upgraded to broken v2.6.0
 var _migrateLocalStorageDataToCookie = function(scope) {
-  var cookieData = Cookie.get(scope.options.cookieName);
+  var cookieData = scope.cookieStorage.get(scope.options.cookieName);
   if (cookieData && cookieData.deviceId) {
     return; // migration not needed
   }
@@ -265,7 +270,7 @@ var _migrateLocalStorageDataToCookie = function(scope) {
     localStorageOptOut = String(localStorageOptOut) === 'true'; // convert to boolean
   }
 
-  Cookie.set(scope.options.cookieName, {
+  scope.cookieStorage.set(scope.options.cookieName, {
     deviceId: cookieDeviceId || localStorageDeviceId,
     userId: cookieUserId || localStorageUserId,
     optOut: (cookieOptOut !== undefined && cookieOptOut !== null) ? cookieOptOut : localStorageOptOut
@@ -273,7 +278,7 @@ var _migrateLocalStorageDataToCookie = function(scope) {
 };
 
 var _loadCookieData = function(scope) {
-  var cookieData = Cookie.get(scope.options.cookieName);
+  var cookieData = scope.cookieStorage.get(scope.options.cookieName);
   if (cookieData) {
     if (cookieData.deviceId) {
       scope.options.deviceId = cookieData.deviceId;
@@ -284,15 +289,44 @@ var _loadCookieData = function(scope) {
     if (cookieData.optOut !== null && cookieData.optOut !== undefined) {
       scope.options.optOut = cookieData.optOut;
     }
+    if (cookieData.sessionId) {
+      scope._sessionId = parseInt(cookieData.sessionId);
+    }
+    if (cookieData.lastEventTime) {
+      scope._lastEventTime = parseInt(cookieData.lastEventTime);
+    }
+    if (cookieData.eventId) {
+      scope._eventId = parseInt(cookieData.eventId);
+    }
+    if (cookieData.identifyId) {
+      scope._identifyId = parseInt(cookieData.identifyId);
+    }
+    if (cookieData.sequenceNumber) {
+      scope._sequenceNumber = parseInt(cookieData.sequenceNumber);
+    }
   }
 };
 
 var _saveCookieData = function(scope) {
-  Cookie.set(scope.options.cookieName, {
+  scope.cookieStorage.set(scope.options.cookieName, {
     deviceId: scope.options.deviceId,
     userId: scope.options.userId,
-    optOut: scope.options.optOut
+    optOut: scope.options.optOut,
+    sessionId: scope._sessionId,
+    lastEventTime: scope._lastEventTime,
+    eventId: scope._eventId,
+    identifyId: scope._identifyId,
+    sequenceNumber: scope._sequenceNumber
   });
+};
+
+// the follow fields used to be saved in localStorage, now saved in cookie to support different subdomains
+var _clearSessionAndEventTrackingFromLocalStorage = function() {
+  localStorage.removeItem(LocalStorageKeys.SESSION_ID);
+  localStorage.removeItem(LocalStorageKeys.LAST_EVENT_TIME);
+  localStorage.removeItem(LocalStorageKeys.LAST_EVENT_ID);
+  localStorage.removeItem(LocalStorageKeys.LAST_IDENTIFY_ID);
+  localStorage.removeItem(LocalStorageKeys.LAST_SEQUENCE_NUMBER);
 };
 
 Amplitude._getUtmParam = function(name, query) {
@@ -325,7 +359,7 @@ Amplitude._getUtmData = function(rawCookie, query) {
  */
 Amplitude.prototype._initUtmData = function(queryParams, cookieParams) {
   queryParams = queryParams || location.search;
-  cookieParams = cookieParams || Cookie.get('__utmz');
+  cookieParams = cookieParams || this.cookieStorage.get('__utmz');
   this._utmProperties = Amplitude._getUtmData(cookieParams, queryParams);
 };
 
@@ -379,10 +413,10 @@ Amplitude.prototype.saveEvents = function() {
 
 Amplitude.prototype.setDomain = function(domain) {
   try {
-    Cookie.options({
+    this.cookieStorage.options({
       domain: domain
     });
-    this.options.domain = Cookie.options().domain;
+    this.options.domain = this.cookieStorage.options().domain;
     _loadCookieData(this);
     _saveCookieData(this);
     //log('set domain=' + domain);
@@ -505,19 +539,17 @@ Amplitude.prototype._logEvent = function(eventType, eventProperties, apiProperti
     var eventId;
     if (eventType === IDENTIFY_EVENT) {
       eventId = this.nextIdentifyId();
-      localStorage.setItem(LocalStorageKeys.LAST_IDENTIFY_ID, eventId);
     } else {
       eventId = this.nextEventId();
-      localStorage.setItem(LocalStorageKeys.LAST_EVENT_ID, eventId);
     }
+    var sequenceNumber = this.nextSequenceNumber();
     var eventTime = new Date().getTime();
     var ua = this._ua;
     if (!this._sessionId || !this._lastEventTime || eventTime - this._lastEventTime > this.options.sessionTimeout) {
       this._sessionId = eventTime;
-      localStorage.setItem(LocalStorageKeys.SESSION_ID, this._sessionId);
     }
     this._lastEventTime = eventTime;
-    localStorage.setItem(LocalStorageKeys.LAST_EVENT_TIME, this._lastEventTime);
+    _saveCookieData(this);
 
     userProperties = userProperties || {};
     // Only add utm properties to user properties for events
@@ -548,7 +580,7 @@ Amplitude.prototype._logEvent = function(eventType, eventProperties, apiProperti
         name: 'amplitude-js',
         version: this.__VERSION__
       },
-      sequence_number: this.nextSequenceNumber() // for ordering events and identifys
+      sequence_number: sequenceNumber // for ordering events and identifys
       // country: null
     };